fix: Address PR review feedback for security framework
Consolidate errors and add performance/reliability improvements:

1. **Error consolidation** - Move PolicyViolation and related exceptions
   to lib/aidp/errors.rb with aliases in Security module for compatibility

2. **Memory leak fix** - SecretsProxy now auto-cleans expired tokens
   every 10 operations to prevent unbounded growth

3. **I/O performance** - SecretsRegistry now caches with file mtime
   checking, avoiding disk reads when cache is valid

4. **Timeout protection** - SecurityCommand audit now runs with 5-minute
   timeout and progress indication

Logging was already present in critical paths (TrifectaState,
RuleOfTwoEnforcer, SecretsProxy). WatchModeHandler already used
hash for O(1) retry count lookup.
